Add experimental add_foreign_utxo to TxBuilder
Partially addresses #141

This is experimetal since it won't work for non-segwit outputs because
we can't set witness_utxo or non_witness_utxo strictly correctly without
knowing whether it's wrapped p2wsh or not and it is difficult to let the
user provide this information without more radical changes.

Also adds the experimental flag.