Add InitWallet structure
Currently when creating a `Wallet` there is a hard requirement that
users call `wallet.sync()` before using the wallet.

In an effort to make the bdk library more intuitive add a new type
`InitWallet` that wraps the `Wallet` and exposes a only three simple
methods. A constructor, a `sync` method, and a `init` method that is a
helper method calling `sync` for simple usage. `sync` consumes `self`
and returns the inner `Wallet`.

Idea based on suggestion by @RCasatta in issue #354.

Fixes #354