Make esplora independent of async-interface
We would like users to be able to extend `bdk` with their own async
blockchain implementations without having to pull in esplora/reqwest.

Do a few manifest improvements and remove the automatic dependency on
`eslpora`/`reqwest` for `async-interface`.

- Remove the dependency from `async-interface` that automatically brings
in `esplora` and `reqwest`.
- Since we do not order dependencies alphabetically; put the `esplora`
dependency next to the `reqwest` dependency to help imply the relation.
- Add a compile conditional to assert that `comapact_filters` is not
enabled at the same time as `async-interface` (as we do for the other
blocking blockchain implementations).