Enable testing of esplora/reqwest
Currently testing of esplora/reqwest is broken because of incorrect
usage of, or lack of, `async`.

Enable testing of esplora/reqwest by using `maybe_await!` macro in tests
and declaring all unit tests in `blockchain_tests.rs` to be async. This
adds a dev dependency on `tokio`.

With this patch applied one can successfully run:

 cargo test --features=test-esplora,use-esplora-reqwest --no-default-features