Handle import error code -4 for `RpcBlockchain`
Bitcoin Core returns code -4 during scriptPubKey/descriptor import
when it is still rescanning the wallet. This PR waits until the
rescan completes before continuing sync as if it was a successful
import.

The caveat of this approach is that a sync cycle may miss out on
importing some scriptPubKeys. Also, re-import logic is not ideal as
network interruptions or parallel requests will result in strange
behavior.

TODO: This may not be this best solution.