Ignore importdescriptor error
The importdescriptor call will throw a `Resource temporarily unavialble`
error only in the first initial sync, where the wallet
has to rescan from genesis.

The sync process carries on as usual after the error, so this patch
makes it ignore only that particular error and carry on the with sync.
All other errors are captured as usual.