[persist_redesign] Split `PersistBackend` into two traits
`PersistBackend` is now two traits:

* `PersistBackend` has method `write_changes()`.
* `LoadablePersistBackend` has method `load_into_tracker()`.

This is because `LoadablePersistBackend` requires a second generic `T`
(for the tracker) which `Store` cannot implement for without introducing
a new tracker trait, or having a concrete type for `T`.

By having two traits, `Store` can still implement `PersistBackend` and
the `LoadablePersistBackend` can be implemented individually with each
concrete tracker type.