[persist_redesign] Modify persist traits
Instead of splitting `PersistBackend` into two traits, we introduce
`Loadable` which is the mimimum functionality for a tracker to be able
to be "loaded into" by the `PersistBackend`.

`Loadable` also makes it easier to make trakers "composable". Typically,
trackers would wrap other tracker implementations internally, and so the
associated changeset would also be wrapped. Having a `Loadable` trait
would allow the new tracker implementation to also have a corresponding
`PersistBackend`.