[examples_redesign] Add best-chain representation to `Tracker`
Introduce field `Tracker::chain` which represents the best-chain
history. It can either be `LocalChain` or `RemoteChain<O: ChainOracle>`.

`Tracker::chain` implements `Loadable`, thus allowing persistance no
matter the implementation. For `RemoteChain`s, we would want to persist
the last-seen height so we know which height to starting syncing from
for the next scan.