`bdk_esplora` redesign
The old logic cannot guarantee consistency of the chain history (i.e. we
may result in a `LocalChain` state which contain blocks that cannot
belong in the same chain). We also had to do a rescan if the tip has
changed since the start of the scan.

To fix this, we anchor transactions at the confirmation height and hash
(which is provided by the esplora API). We fetch the `LocalChain` update
separately, and guarantee consistency by fetching most-recent blocks
atomically via the `GET /blocks` endpoint.

Because the anchors need to be reflected in `LocalChain`,
`TxGraph::missing_blocks(&self, chain: &LocalChain)` is introduced to
fetch missing blocks in a separate call.