Create CBFUpdate iterator and convert CBFUpdate into TxGraph update
CBFUpdateIterator will return CBFUpdate data until the Nakamoto client
has reached the tip of the chain. At that point, the iterator will
return None.

The new method, `into_tx_graph_update`, will convert the CBFUpdate into
a TxGraph update. This method is a helper that will be necessary for
updating the TxGraph once we have a CBFUpdate.