feat: Update into_tx_graph_update and add syncing
Update into_tx_graph_update to take a Vec of tuples of
blocks and transactions as input to create a TxGraph.

Add scanning with Nakamoto client. We start by creating a list of scripts
to scan and we continue to scan the chain until we have satisfied
stop_gap condition. On each scan we get a fresh list of script to watch.
We then update the IndexedGraph and return the IndexedAdditions.