test(wallet): test anti fee sniping logic
This change adds two tests to wallet.rs. In
`test_create_tx_fee_sniping_tr` we check that for a taproot
transaction with rbf enabled, either the tx LockTime is set to the most
recent checkpoint, or the input's Sequence is set to its confirmations
(allowing for randomness in both cases).

In `test_create_tx_fee_sniping_tr_csv` we check that policy
csv requirements take precedence over a default nSequence. A new
test descriptor getter is added to common.rs
`get_test_tr_with_taptree_csv` that contains a script path with a csv
condition.